backend/requirements.txt
speedtest-cli
flask
flask-cors
schedule